+++ CHANGELOG.md
@@ -2,4 +2,5 @@
 
 ## Version 0.2.0
 
-- `get-names` and `get-deprecated-names` functions.
+- `get-names`, `get-deprecated-names`, and `get-math-names` functions.
+- `codex-version` constant.